done some refactoring
