+++ ChartWRCLibrary/ChartWRCLibrary.vcxproj
@@ -37,6 +37,7 @@
     <WindowsTargetPlatformVersion>10.0.16299.0</WindowsTargetPlatformVersion>
     <WindowsTargetPlatformMinVersion>10.0.16299.0</WindowsTargetPlatformMinVersion>
     <ApplicationTypeRevision>10.0</ApplicationTypeRevision>
+    <ProjectName>ChartWRCLibrary</ProjectName>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -206,11 +207,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="DelegateCommand.hpp" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="FftHelper.hpp" />
     <ClInclude Include="Source\FftRealPair.hpp" />
+    <ClInclude Include="TestConverter.hpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="BindableBase.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
@@ -221,6 +225,7 @@
     </ClCompile>
     <ClCompile Include="FftHelper.cpp" />
     <ClCompile Include="Source\FftRealPair.cpp" />
+    <ClCompile Include="RatingConverter.hpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
